--- JoIsElDa/client/src/components/main.js
@@ -1,22 +1,0 @@
-import React, { Component } from "react";
-import { Routes, Route } from "react-router-dom";
-
-
-
-const Main = () => {
-
-
-    return (
-        <div>
-            <Routes>
-                
-                
-            </Routes>
-
-        </div>
-
-    );
-
-}
-
-export default Main;
