--- JoIsElDa/package.json
@@ -1,1 +1,0 @@
-{}
